+++ src/celeritas/em/interactor/AtomicRelaxation.hh
@@ -11,8 +11,8 @@
 #include "corecel/Types.hh"
 #include "corecel/cont/MiniStack.hh"
 #include "corecel/cont/Span.hh"
+#include "corecel/random/distribution/IsotropicDistribution.hh"
 #include "corecel/random/distribution/Selector.hh"
-#include "geocel/random/IsotropicDistribution.hh"
 #include "celeritas/Quantities.hh"
 #include "celeritas/Types.hh"
 #include "celeritas/em/data/AtomicRelaxationData.hh"
