+++ src/celeritas/em/interactor/AtomicRelaxationHelper.hh
@@ -9,8 +9,8 @@
 #include "corecel/Macros.hh"
 #include "corecel/Types.hh"
 #include "corecel/cont/Span.hh"
+#include "corecel/random/distribution/IsotropicDistribution.hh"
 #include "corecel/sys/ThreadId.hh"
-#include "geocel/random/IsotropicDistribution.hh"
 #include "celeritas/Quantities.hh"
 #include "celeritas/em/data/AtomicRelaxationData.hh"
