+++ src/celeritas/optical/interactor/WavelengthShiftGenerator.hh
@@ -10,8 +10,8 @@
 #include "corecel/Types.hh"
 #include "corecel/data/StackAllocator.hh"
 #include "corecel/random/distribution/ExponentialDistribution.hh"
+#include "corecel/random/distribution/IsotropicDistribution.hh"
 #include "corecel/random/distribution/UniformRealDistribution.hh"
-#include "geocel/random/IsotropicDistribution.hh"
 #include "celeritas/Types.hh"
 #include "celeritas/grid/NonuniformGridCalculator.hh"
 #include "celeritas/optical/Interaction.hh"
