+++ test/celeritas/geo/HeuristicGeoExecutor.hh
@@ -14,12 +14,12 @@
 #include "corecel/math/ArrayUtils.hh"
 #include "corecel/math/Atomics.hh"
 #include "corecel/random/distribution/BernoulliDistribution.hh"
+#include "corecel/random/distribution/IsotropicDistribution.hh"
+#include "corecel/random/distribution/UniformBoxDistribution.hh"
 #include "corecel/random/distribution/UniformRealDistribution.hh"
 #include "corecel/random/engine/RngEngine.hh"
 #include "corecel/sys/ThreadId.hh"
 #include "geocel/UnitUtils.hh"
-#include "geocel/random/IsotropicDistribution.hh"
-#include "geocel/random/UniformBoxDistribution.hh"
 #include "celeritas/Units.hh"
 #include "celeritas/geo/GeoTrackView.hh"
