+++ test/celeritas/global/StepperGeant.test.cc
@@ -9,9 +9,9 @@
 #include "corecel/Types.hh"
 #include "corecel/cont/Range.hh"
 #include "corecel/cont/Span.hh"
+#include "corecel/random/distribution/IsotropicDistribution.hh"
 #include "corecel/sys/ActionRegistry.hh"
 #include "geocel/UnitUtils.hh"
-#include "geocel/random/IsotropicDistribution.hh"
 #include "celeritas/LeadBoxTestBase.hh"
 #include "celeritas/OneSteelSphereBase.hh"
 #include "celeritas/TestEm15Base.hh"
