+++ test/celeritas/optical/OpticalCollector.test.cc
@@ -17,9 +17,9 @@
 #include "corecel/io/LogContextException.hh"
 #include "corecel/io/Logger.hh"
 #include "corecel/math/Algorithms.hh"
+#include "corecel/random/distribution/IsotropicDistribution.hh"
 #include "corecel/sys/ActionRegistry.hh"
 #include "geocel/UnitUtils.hh"
-#include "geocel/random/IsotropicDistribution.hh"
 #include "celeritas/LArSphereBase.hh"
 #include "celeritas/alongstep/AlongStepUniformMscAction.hh"
 #include "celeritas/em/params/UrbanMscParams.hh"
