+++ test/corecel/CMakeLists.txt
@@ -108,12 +108,14 @@ celeritas_add_test(random/distribution/BernoulliDistribution.test.cc)
 celeritas_add_test(random/distribution/ExponentialDistribution.test.cc)
 celeritas_add_test(random/distribution/GammaDistribution.test.cc)
 celeritas_add_test(random/distribution/InverseSquareDistribution.test.cc)
+celeritas_add_test(random/distribution/IsotropicDistribution.test.cc)
 celeritas_add_test(random/distribution/NormalDistribution.test.cc)
 celeritas_add_test(random/distribution/PoissonDistribution.test.cc)
 celeritas_add_test(random/distribution/PowerDistribution.test.cc)
 celeritas_add_test(random/distribution/RadialDistribution.test.cc)
 celeritas_add_test(random/distribution/ReciprocalDistribution.test.cc)
 celeritas_add_test(random/distribution/RejectionSampler.test.cc)
+celeritas_add_test(random/distribution/UniformBoxDistribution.test.cc)
 celeritas_add_test(random/distribution/UniformRealDistribution.test.cc)
 celeritas_add_test(random/distribution/DistributionTypeTraits.test.cc GPU)
