+++ test/orange/univ/RectArrayTracker.test.cc
@@ -16,9 +16,9 @@
 #include "corecel/data/Ref.hh"
 #include "corecel/io/Repr.hh"
 #include "corecel/math/ArrayUtils.hh"
+#include "corecel/random/distribution/IsotropicDistribution.hh"
+#include "corecel/random/distribution/UniformBoxDistribution.hh"
 #include "corecel/sys/Stopwatch.hh"
-#include "geocel/random/IsotropicDistribution.hh"
-#include "geocel/random/UniformBoxDistribution.hh"
 #include "orange/OrangeGeoTestBase.hh"
 #include "orange/OrangeParams.hh"
 #include "orange/detail/UniverseIndexer.hh"
